updating branch

+++ modules/drivers/canbus/can_comm/message_manager.h
@@ -196,7 +196,8 @@ void MessageManager<SensorType>::Parse(const uint32_t message_id,
     it->second.real_period = time - it->second.last_time;
     // if period 1.5 large than base period, inc error_count
     const double period_multiplier = 1.5;
-    if (it->second.real_period > (it->second.period * period_multiplier)) {
+    if (static_cast<double>(it->second.real_period) >
+        (static_cast<double>(it->second.period) * period_multiplier)) {
       it->second.error_count += 1;
     } else {
       it->second.error_count = 0;
